sim1750/Makefile - At rule `debug' add gcc switch -Og. - At rule `clean' use $(RM) in lieu of `rm'.
sim1750/src/type.h
- Move definition of `uchar' outside #ifndef HAVE_UTYPES.

sim1750/src/tekhex.h
- #include "type.h".

sim1750/src/tekops.c
- In function si_lo initialize retval.

as1750/Makefile
- Derive OBJ from SRC. Remove redundant rule.

as1750/common.h
as1750/as1750.h
as1750/as1750.c
as1750/flt1750.c
as1750/syms.h
as1750/syms.c
as1750/tldldm.h
as1750/tldldm.c
as1750/utils.c
as1750/main.c
- Replace usage of long/ulong by int/uint.
  Adjust hex constants and printf formats to this change.
- Remove #ifdefs for supporting ASL (as1750 has been integrated there).

as1750/type.h
as1750/tekhex.h
as1750/tekhex.c
- Copy from sim1750.

+++ as1750/common.h
@@ -50,8 +50,8 @@ struct objblock {
 	 int       n_used;
 	 char      *name;
 	 section_t section;
-	 long      start_address;
-/* when start_address is -1L, use default linking mechanism (see main.c):
+	 int       start_address;
+/* when start_address is -1 use default linking mechanism (see main.c):
    Init   at 0x00100
    Normal at Init + sizeof(Init)
    Konst  at Normal + sizeof(Normal)
