+++ as1750/tekhex.h
@@ -3,8 +3,8 @@
 #include "type.h"
 
 extern int      check_tekline (char *line);
-extern void     emit_tekword (ulong startaddr, unsigned short word);
+extern void     emit_tekword (uint startaddr, unsigned short word);
 extern void     finish_tekline ();
-extern unsigned create_tekfile (char *outfname);
-extern void     close_tekfile (ulong transfer_address);
+extern uint     create_tekfile (char *outfname);
+extern void     close_tekfile (uint transfer_address);
