+++ sim1750/src/tekhex.h
@@ -1,8 +1,10 @@
 /* tekhex.h  --  exports of tekhex.c */
 
-extern int    check_tekline (char *line);
-extern void   emit_tekword (unsigned long startaddr, unsigned short word);
-extern void   finish_tekline ();
-extern int    create_tekfile (char *outfname);
-extern void   close_tekfile (unsigned long transfer_address);
+#include "type.h"
+
+extern int      check_tekline (char *line);
+extern void     emit_tekword (uint startaddr, unsigned short word);
+extern void     finish_tekline ();
+extern uint     create_tekfile (char *outfname);
+extern void     close_tekfile (uint transfer_address);
